rename(#12): 코드 컨벤션에 맞춰 디렉토리 명 수정
dto 디렉토리를 model 디렉토리로 변경하였습니다.
