+++ server/src/main/java/com/hexacore/tayo/test/model/TestDto.java
@@ -1,4 +1,4 @@
-package com.hexacore.tayo.test.dto;
+package com.hexacore.tayo.test.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,5 +6,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class TestDto {
+
     private String text;
 }
